requirements.txt
-i https://pypi.org/simple
bcrypt==4.0.1; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
flash==1.0.3
flask==2.2.2; python_version >= '3.7'
flask-bcrypt==1.0.1
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
pymysql==1.0.2
werkzeug==2.2.2; python_version >= '3.7'
